requirements.txt
aiofiles==24.1.0
altair==5.3.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
cryptography==43.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.15.4
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.4
openai==1.3.5
packaging==24.1
pandas==2.0.3
pathlib==1.0.1
pdfminer.six==20240706
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.2
prov==2.0.1
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.2
pdfplumber=0.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.3
scipy==1.10.1
sentencepiece==0.2.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.1
streamlit==1.36.0
sympy==1.13.1
tbb==2021.13.0
tenacity==8.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traits==6.3.2
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
watchdog==4.0.1
zipp==3.19.2
